+++ core/src/script/watch.ts
@@ -1,6 +1,4 @@
-
-
-import createCompiler from "../utils"
+import createCompiler from "../utils/getWebpackConfig"
 import clearConsole from 'react-dev-utils/clearConsole';
 import { OptionsProps } from "../interface"
 import { webpackConfigPath } from "./../overrides/pathUtils"
