+++ core/src/utils/output/index.ts
@@ -1,0 +1,12 @@
+import { WebpackConfiguration } from 'kkt';
+
+/** 输出配置 */
+export const restOutPut = (conf: WebpackConfiguration, options: WebpackConfiguration['output']): WebpackConfiguration => {
+  return {
+    ...conf,
+    output: {
+      ...conf.output,
+      ...options,
+    },
+  };
+};
